refactor: Restructure backend modules and remove unused components
- Removed AuthController, ExpenseController, and related DTOs and services to streamline the backend.
- Updated AuthModule and ExpenseModule to reflect new file structure.
- Enhanced main.ts to utilize Logger for better logging practices.
- Revised README to improve project architecture documentation.
