+++ apps/backend/src/app/auth/user.schema.ts
