+++ apps/backend/src/main.ts
@@ -4,11 +4,12 @@
  */
 
 import { NestFactory } from '@nestjs/core';
-import { ValidationPipe } from '@nestjs/common';
+import { Logger, ValidationPipe } from '@nestjs/common';
 import { AppModule } from './app/app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
+  const logger = new Logger('Bootstrap');
   
   // Enable CORS
   app.enableCors({
@@ -28,7 +29,7 @@ async function bootstrap() {
   
   const port = process.env.PORT || 3001;
   await app.listen(port);
-  console.log(`Application is running on: http://localhost:${port}`);
-  console.log(`MongoDB URI: ${process.env.MONGODB_URI || 'mongodb://localhost:27017/expense-tracker'}`);
+  logger.log(`Application is running on: http://localhost:${port}`);
+  logger.log(`MongoDB URI: ${process.env.MONGODB_URI || 'mongodb://localhost:27017/expense-tracker'}`);
 }
 bootstrap();
